backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
python-dotenv==1.0.1
pydantic==2.10.4
aiosqlite==0.20.0
fpdf2==2.8.2
